backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
neo4j==5.15.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
python-multipart==0.0.7
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
kubernetes==28.1.0
asyncio==3.4.3
psutil==5.9.7
mysql-connector-python==8.2.0 